Add cache using redis

+++ src/services/postgres/PlaylistsService.js
@@ -2,10 +2,12 @@ const { nanoid } = require('nanoid');
 const { Pool } = require('pg');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 const BadRequestError = require('../../exceptions/BadRequestError');
+const StringUtils = require('../../utils/StringUtils');
 
 class PlaylistsService {
-  constructor(collabService) {
+  constructor(collabService, cacheService) {
     this._pool = new Pool();
+    this._cacheService = cacheService;
     this._collabService = collabService;
   }
 
@@ -21,32 +23,44 @@ class PlaylistsService {
       return new BadRequestError('Failed when adding new playlist');
     }
 
+    await this._deletePlaylistCache(ownerUserId);
     return rows[0].id;
   }
 
   async getPlaylistsByUserId(userId) {
-    const query = {
-      text: `SELECT playlists.id, name, username FROM playlists 
-            LEFT JOIN users ON users.id = playlists.owner 
-            LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id 
-            WHERE owner = $1 OR collaborations.user_id = $1 GROUP BY playlists.id, name, username`,
-      values: [userId],
-    };
+    const cacheKey = StringUtils.getPlaylistCacheKey(userId);
 
-    const { rows } = await this._pool.query(query);
-    return rows;
+    try {
+      const playlists = await this._cacheService.get(cacheKey);
+      return JSON.parse(playlists);
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, name, username FROM playlists 
+                LEFT JOIN users ON users.id = playlists.owner 
+                LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id 
+                WHERE owner = $1 OR collaborations.user_id = $1 GROUP BY playlists.id, name, username`,
+        values: [userId],
+      };
+
+      const { rows } = await this._pool.query(query);
+
+      await this._cacheService.set(cacheKey, JSON.stringify(rows));
+      return rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1',
+      text: 'DELETE FROM playlists WHERE id = $1 returning owner',
       values: [id],
     };
 
-    const { rowCount } = await this._pool.query(query);
-    if (!rowCount) {
+    const { rows } = await this._pool.query(query);
+    if (!rows.length) {
       throw new BadRequestError('Failed when deleting playlist');
     }
+
+    await this._deletePlaylistCache(rows[0].owner);
   }
 
   async verifyPlaylistOwner(playlistId, userId) {
@@ -83,6 +97,8 @@ class PlaylistsService {
     if (!rowCount) {
       throw new BadRequestError('Failed when adding song to the playlist.');
     }
+
+    await this._deletePlaylistSongCache(playlistId);
   }
 
   async deleteSongFromPlaylist(playlistId, songId) {
@@ -95,6 +111,18 @@ class PlaylistsService {
     if (!rowCount) {
       throw new BadRequestError('Failed when deleting song from playlist');
     }
+
+    await this._deletePlaylistSongCache(playlistId);
+  }
+
+  async _deletePlaylistCache(ownerUserId) {
+    const cacheKey = StringUtils.getPlaylistCacheKey(ownerUserId);
+    await this._cacheService.delete(cacheKey);
+  }
+
+  async _deletePlaylistSongCache(playlistId) {
+    const cacheKey = StringUtils.getPlaylistSongCacheKey(playlistId);
+    await this._cacheService.delete(cacheKey);
   }
 }
